Display number of Likes for each item
