+++ webpack.config.js
@@ -6,7 +6,6 @@ module.exports = {
   entry: {
     index: './src/index.js'
   },
-  devtool: 'inline-source-map',
   devServer: {
     contentBase: './dist',
   },
